This folder contains a ladder program with timer and couter function blocks
Controls a motor wich moves a conveior and count pieces. It DOESN´T considers differences between materials (metalic or not). Just count number of pieces and time. If the time exceeds 10s, the motor is turned off; if 5 pieces are counted, the same; or a b0 button pressed, also turns off the motor.